+++ python/python/tests/test_async_client.py
@@ -7269,11 +7269,18 @@ class TestCommands:
         value2 = get_random_string(5)
 
         assert await redis_client.set(key, value) == OK
-        assert await redis_client.wait(1, 1000) >= 0
+        if isinstance(redis_client, GlideClusterClient):
+            assert await redis_client.wait(1, 1000) >= 1
+        else:
+            assert await redis_client.wait(1, 1000) >= 0
 
         # ensure that command doesn't time out even if timeout > request timeout (250ms by default)
         assert await redis_client.set(key, value2) == OK
         assert await redis_client.wait(100, 500) >= 0
+
+        # command should fail on a negative timeout value
+        with pytest.raises(RequestError):
+            await redis_client.wait(1, -1)
 
     @pytest.mark.parametrize("cluster_mode", [True, False])
     @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])